+++ packages/velodyne-cloud/src/Calibration.ts
@@ -2,16 +2,11 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
+import { CalibrationData, loadCalibrationData } from "./CalibrationData";
 import { LaserCorrection, Model } from "./VelodyneTypes";
-import HDL32E_json from "./data/calibration/32db.json";
-import HDL64E_S21_json from "./data/calibration/64e_s2.1-sztaki.json";
-import HDL64E_S3_json from "./data/calibration/64e_s3-xiesc.json";
-import HDL64E_json from "./data/calibration/64e_utexas.json";
-import VLP16HiRes_json from "./data/calibration/VLP16_hires_db.json";
-import VLP16_json from "./data/calibration/VLP16db.json";
-import VLS128_json from "./data/calibration/VLS128.json";
-import VLP32C_json from "./data/calibration/VeloView-VLP-32C.json";
 
+// Takes a calibration data file as input and computes cached lookup tables for
+// use in online point cloud conversion
 export class Calibration {
   static ROTATION_RESOLUTION = 0.01; // [deg]
   static ROTATION_MAX_UNITS = 36000; // [deg/100]
@@ -30,10 +25,10 @@ export class Calibration {
   readonly cosRotTable: number[];
   readonly vls128LaserAzimuthCache: number[];
 
-  constructor(model: Model) {
+  constructor(model: Model, calibrationData?: CalibrationData) {
     this.model = model;
-    const data = Calibration.Data(model);
-    this.laserCorrections = data.lasers.map((v) => {
+    calibrationData = calibrationData ?? loadCalibrationData(model);
+    this.laserCorrections = calibrationData.lasers.map((v) => {
       return {
         laserId: v.laser_id,
         rotCorrection: v.rot_correction,
@@ -54,7 +49,7 @@ export class Calibration {
         sinVertCorrection: Math.sin(v.vert_correction),
       };
     });
-    this.distanceResolution = data.distance_resolution;
+    this.distanceResolution = calibrationData.distance_resolution;
     this.timingOffsets = Calibration.BuildTimingsFor(model);
 
     // Set up cached values for sin and cos of all the possible headings
@@ -123,51 +118,9 @@ export class Calibration {
           .map((_col, y) => fullFiring * block(x, y) + singleFiring * point(x, y) + offset),
       );
   }
-
-  static Data(model: Model): CalibrationData {
-    switch (model) {
-      case Model.VLP16:
-        return VLP16_json as CalibrationData;
-      case Model.VLP16HiRes:
-        return VLP16HiRes_json as CalibrationData;
-      case Model.VLP32C:
-        return VLP32C_json as CalibrationData;
-      case Model.HDL32E:
-        return HDL32E_json as CalibrationData;
-      case Model.HDL64E:
-        return HDL64E_json as CalibrationData;
-      case Model.HDL64E_S21:
-        return HDL64E_S21_json as CalibrationData;
-      case Model.HDL64E_S3:
-        return HDL64E_S3_json as CalibrationData;
-      case Model.VLS128:
-        return VLS128_json as CalibrationData;
-    }
-  }
 }
 
 type IndexCalc = (x: number, y: number) => number;
-
-type LaserEntry = {
-  rot_correction: number;
-  vert_correction: number;
-  dist_correction: number;
-  two_pt_correction_available?: boolean;
-  dist_correction_x: number;
-  dist_correction_y: number;
-  vert_offset_correction: number;
-  horiz_offset_correction: number;
-  max_intensity?: number;
-  min_intensity?: number;
-  focal_distance: number;
-  focal_slope: number;
-  laser_id: number;
-};
-
-type CalibrationData = {
-  lasers: LaserEntry[];
-  distance_resolution: number;
-};
 
 function deg2rad(degrees: number): number {
   return degrees * (Math.PI / 180);
